# Tools: remove duplicated parameters to be setted back