Example on Copter.ThrottleFailsafe,
we pass from :
`
[('DISARM_DELAY', 10.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('DISARM_DELAY', 0.0), ('FS_THR_ENABLE', 1.0), ('FS_OPTIONS', 0.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_THR_ENABLE', 0.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_THR_ENABLE', 1.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_THR_ENABLE', 3.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_THR_ENABLE', 4.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_THR_ENABLE', 5.0), ('SIM_GPS_DISABLE', 0.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('SIM_GPS_DISABLE', 1.0), ('FS_THR_ENABLE', 1.0), ('SIM_GPS_DISABLE', 0.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('SIM_GPS_DISABLE', 1.0), ('FS_THR_ENABLE', 4.0), ('SIM_GPS_DISABLE', 0.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('SIM_GPS_DISABLE', 1.0), ('FS_THR_ENABLE', 5.0), ('SIM_GPS_DISABLE', 0.0), ('SIM_GPS_DISABLE', 1.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_THR_ENABLE', 4.0), ('SIM_GPS_DISABLE', 0.0), ('SIM_GPS_DISABLE', 1.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('SIM_SPEEDUP', 8.0), ('FS_GCS_ENABLE', 0.0), ('FS_THR_ENABLE', 5.0), ('FS_OPTIONS', 0.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_GCS_ENABLE', 1.0), ('SIM_SPEEDUP', 4.0), ('FS_OPTIONS', 4.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_OPTIONS', 1.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('FS_THR_ENABLE', 1.0)]`

to 
`
[('DISARM_DELAY', 10.0), ('SIM_RC_FAIL', 0.0), ('SIM_RC_FAIL', 1.0), ('DISARM_DELAY', 0.0), ('FS_THR_ENABLE', 1.0), ('FS_OPTIONS', 0.0), ('FS_THR_ENABLE', 0.0), ('FS_THR_ENABLE', 3.0), ('FS_THR_ENABLE', 4.0), ('FS_THR_ENABLE', 5.0), ('SIM_GPS_DISABLE', 0.0), ('SIM_GPS_DISABLE', 1.0), ('SIM_SPEEDUP', 8.0), ('FS_GCS_ENABLE', 0.0), ('FS_GCS_ENABLE', 1.0), ('SIM_SPEEDUP', 4.0), ('FS_OPTIONS', 4.0), ('FS_OPTIONS', 1.0)]`

This save us a large amount of time : 
Master : real 138,58s user 42,47s system 88% cpu 3:25,12 total
PR : real 128,33s user 38,58s system 88% cpu 3:08,02 total