@khancyr 2020-08-11
Question : should we remove completely all duplicate parameter by name ? It would save us a lot more time and I don't think we need those duplicate