@peterbarker 2020-08-11
Might be sufficient to not push onto the stack if we've already got the key on there.

Care may need to be taken around enable parameters - but probably not.