@khancyr 2020-08-12
changed to not push duplicate in the list.
for Copter.ThrottleFailsafe, it gives now

[('DISARM_DELAY', 10.0), ('SIM_RC_FAIL', 0.0), ('FS_THR_ENABLE', 1.0), ('FS_OPTIONS', 0.0), ('SIM_GPS_DISABLE', 0.0), ('SIM_SPEEDUP', 8.0), ('FS_GCS_ENABLE', 0.0)]